chore(utils/__tests__): move assert-resolvable.test.ts from the bundler packages
